Move generic his components to general
